+++ commands/operator-sdk/cmd/add/controller.go
@@ -15,6 +15,8 @@
 package add
 
 import (
+	"fmt"
+
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
 	"github.com/operator-framework/operator-sdk/pkg/scaffold"
 	"github.com/operator-framework/operator-sdk/pkg/scaffold/input"
@@ -24,7 +26,7 @@ import (
 )
 
 func NewControllerCmd() *cobra.Command {
-	apiCmd := &cobra.Command{
+	controllerCmd := &cobra.Command{
 		Use:   "controller",
 		Short: "Adds a new controller pkg",
 		Long: `operator-sdk add controller --kind=<kind> --api-version=<group/version> creates a new
@@ -44,29 +46,35 @@ Example:
 	└── controller.go
 
 `,
-		Run: controllerRun,
+		RunE: controllerRun,
 	}
 
-	apiCmd.Flags().StringVar(&apiVersion, "api-version", "", "Kubernetes APIVersion that has a format of $GROUP_NAME/$VERSION (e.g app.example.com/v1alpha1)")
-	apiCmd.MarkFlagRequired("api-version")
-	apiCmd.Flags().StringVar(&kind, "kind", "", "Kubernetes resource Kind name. (e.g AppService)")
-	apiCmd.MarkFlagRequired("kind")
+	controllerCmd.Flags().StringVar(&apiVersion, "api-version", "", "Kubernetes APIVersion that has a format of $GROUP_NAME/$VERSION (e.g app.example.com/v1alpha1)")
+	if err := controllerCmd.MarkFlagRequired("api-version"); err != nil {
+		log.Fatalf("Failed to mark `api-version` flag for `add controller` subcommand as required")
+	}
+	controllerCmd.Flags().StringVar(&kind, "kind", "", "Kubernetes resource Kind name. (e.g AppService)")
+	if err := controllerCmd.MarkFlagRequired("kind"); err != nil {
+		log.Fatalf("Failed to mark `kind` flag for `add controller` subcommand as required")
+	}
 
-	return apiCmd
+	return controllerCmd
 }
 
-func controllerRun(cmd *cobra.Command, args []string) {
-	// Only Go projects can add controllers.
-	projutil.MustGoProjectCmd(cmd)
-
+func controllerRun(cmd *cobra.Command, args []string) error {
 	projutil.MustInProjectRoot()
+
+	// Only Go projects can add controllers.
+	if err := projutil.CheckGoProjectCmd(cmd); err != nil {
+		return err
+	}
 
 	log.Infof("Generating controller version %s for kind %s.", apiVersion, kind)
 
 	// Create and validate new resource
 	r, err := scaffold.NewResource(apiVersion, kind)
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	cfg := &input.Config{
@@ -80,8 +88,9 @@ func controllerRun(cmd *cobra.Command, args []string) {
 		&scaffold.AddController{Resource: r},
 	)
 	if err != nil {
-		log.Fatalf("Add scaffold failed: (%v)", err)
+		return fmt.Errorf("controller scaffold failed: (%v)", err)
 	}
 
 	log.Info("Controller generation complete.")
+	return nil
 }