+++ commands/operator-sdk/cmd/root.go
@@ -34,9 +34,11 @@ func NewRootCmd() *cobra.Command {
 	cmd.AddCommand(NewUpCmd())
 	cmd.AddCommand(NewCompletionCmd())
 	cmd.AddCommand(NewTestCmd())
+	cmd.AddCommand(NewScorecardCmd())
 	cmd.AddCommand(NewPrintDepsCmd())
 	cmd.AddCommand(NewMigrateCmd())
 	cmd.AddCommand(NewRunCmd())
+	cmd.AddCommand(NewOLMCatalogCmd())
 
 	return cmd
 }