+++ pkg/scaffold/ansible/deploy_operator.go
@@ -57,9 +57,6 @@ spec:
         - name: {{.ProjectName}}
           # Replace this with the built image name
           image: "{{ "{{ REPLACE_IMAGE }}" }}"
-          ports:
-          - containerPort: 60000
-            name: metrics
           imagePullPolicy: "{{ "{{ pull_policy|default('Always') }}"}}"
           env:
             - name: WATCH_NAMESPACE
@@ -70,6 +67,10 @@ spec:
                 fieldRef:
                   fieldPath: metadata.namespace
               {{- end}}
+            - name: POD_NAME
+              valueFrom:
+                fieldRef:
+                  fieldPath: metadata.name
             - name: OPERATOR_NAME
               value: "{{.ProjectName}}"
 `