+++ pkg/scaffold/build_dockerfile_test.go
@@ -35,9 +35,17 @@ func TestDockerfile(t *testing.T) {
 
 const dockerfileExp = `FROM alpine:3.8
 
-RUN apk upgrade --update --no-cache
+ENV OPERATOR=/usr/local/bin/app-operator \
+    USER_UID=1001 \
+    USER_NAME=app-operator
 
-USER nobody
+# install operator binary
+COPY build/_output/bin/app-operator ${OPERATOR}
 
-ADD build/_output/bin/app-operator /usr/local/bin/app-operator
+COPY build/bin /usr/local/bin
+RUN  /usr/local/bin/user_setup
+
+ENTRYPOINT ["/usr/local/bin/entrypoint"]
+
+USER ${USER_UID}
 `