+++ pkg/scaffold/cmd.go
@@ -48,7 +48,7 @@ import (
 
 	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 	"github.com/operator-framework/operator-sdk/pkg/leader"
-	"github.com/operator-framework/operator-sdk/pkg/ready"
+	"github.com/operator-framework/operator-sdk/pkg/metrics"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
@@ -57,6 +57,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/runtime/signals"
 )
 
+// Change below variables to serve metrics on different host or port.
+var (
+    metricsHost = "0.0.0.0"
+    metricsPort int32 = 8383
+)
 var log = logf.Log.WithName("cmd")
 
 func printVersion() {
@@ -88,29 +93,21 @@ func main() {
 		log.Error(err, "")
 		os.Exit(1)
 	}
+	
+	ctx := context.TODO()
 
 	// Become the leader before proceeding
-	err = leader.Become(context.TODO(), "{{ .ProjectName }}-lock")
+	err = leader.Become(ctx, "{{ .ProjectName }}-lock")
 	if err != nil {
 		log.Error(err, "")
 		os.Exit(1)
 	}
-
-	r := ready.NewFileReady()
-	err = r.Set()
-	if err != nil {
-		log.Error(err, "")
-		os.Exit(1)
-	}
-	defer func() {
-		if err = r.Unset(); err != nil {
-			log.Error(err, "")
-			os.Exit(1)
-		}
-	}()
 
 	// Create a new Cmd to provide shared dependencies and start components
-	mgr, err := manager.New(cfg, manager.Options{Namespace: namespace})
+	mgr, err := manager.New(cfg, manager.Options{
+		Namespace:          namespace,
+		MetricsBindAddress: fmt.Sprintf("%s:%d", metricsHost, metricsPort),
+	})	
 	if err != nil {
 		log.Error(err, "")
 		os.Exit(1)
@@ -128,6 +125,12 @@ func main() {
 	if err := controller.AddToManager(mgr); err != nil {
 		log.Error(err, "")
 		os.Exit(1)
+	}
+
+	// Create Service object to expose the metrics port.
+	_, err = metrics.ExposeMetricsPort(ctx, metricsPort)
+	if err != nil {
+		log.Info(err.Error())
 	}
 
 	log.Info("Starting the Cmd.")